+++ banto/src/main/java/com/example/banto/BantoApplication.java
@@ -2,6 +2,7 @@ package com.example.banto;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 public class BantoApplication {
