+++ banto/src/main/java/com/example/banto/Controllers/UserController.java
@@ -1,19 +1,26 @@
 package com.example.banto.Controllers;
 
+import java.net.http.HttpRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import com.example.banto.DTOs.UserDTO;
 import com.example.banto.Services.UserService;
+import com.exmaple.banto.JWTs.JwtUtil;
 
 @Controller
 public class UserController {
 	
 	@Autowired
 	UserService userService;
+	
+	JwtUtil jwtUtil = new JwtUtil();
 	
 	// 회원가입 기능
 	@PostMapping("/sign")
@@ -32,7 +39,8 @@ public class UserController {
 		try {
 			// 단순히 유저 찾기에 불과, JWT 토큰 또는 세션이나 비밀번호 해시화에 대해 넣을 필요가 있음
 			userService.login(dto);
-			return ResponseEntity.ok().body(null);
+			//email이 담긴 JWT 반환
+			return ResponseEntity.ok().body(jwtUtil.generateToken(dto.getEmail()));
 		}catch(Exception e) {
 			return ResponseEntity.badRequest().body(e.getMessage());
 		}
