+++ banto/src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.devtools.restart.enabled=true
 spring.devtools.livereload.enabled=true
 spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
 
-spring.datasource.url=jdbc:oracle:thin:@220.79.17.81:1521:xe
+spring.datasource.url=jdbc:oracle:thin:@localhost:1521:xe
 
 spring.datasource.username=bangtoshop
 spring.datasource.password=1234
